backend/requirements.txt
# Flask core
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==5.0.0
Flask-Migrate==4.0.7

# Database
PyMySQL==1.1.1
cryptography==43.0.0
python-dotenv==1.0.1
bcrypt==4.2.0

# MQTT
paho-mqtt==2.1.0

# Elasticsearch
elasticsearch==8.12.1

# Redis
redis==5.0.8

# WSGI server (production)
waitress==3.0.0

# Report generation
reportlab==4.2.2
openpyxl==3.1.5

# Monitoring
prometheus-flask-exporter==0.23.2
prometheus-client==0.21.0

# Testing
pytest==8.3.2
pytest-cov==5.0.0
